renamed files and more structural improvements
