+++ A2/Part2/Part_II_AjDonald_JayvenLarsen.cpp
@@ -1,3 +1,10 @@
+/*
+* SYSC4001 A2 Part2 
+*
+* Authors: Aj Donald 101259149 & Jayven Larsen 101260364
+*
+*/
+
 #include <iostream>
 using namespace std;
 #include <unistd.h>
