+++ A2/Part2/Part_II_AjDonald_JayvenLarsen_program2.cpp
@@ -1,12 +1,8 @@
 /*
 * SYSC4001 A2 Part2 
 *
-* Authors: Aj Donald 101259149 & Simon Hagos
+* Authors: Aj Donald 101259149 & Jayven Larsen 101260364
 *
-* Note: Due to the difficulties encountered during the assignment of the teams for A2, 
-this assignment had already been completed with my previous partner from A1, (Jayven Larsen, 101260364).
-As a result, the solutions of this submission for all parts was completed with him. Having discussed with Professor Wainer,
- * we have understood that there shouldn't be any risk of plagiarism between our two groups given this mistake. 
 */
 
 #include <iostream>
