--- A2/Part3/test1.sh
@@ -1,7 +1,0 @@
-g++ -std=c++11 interrupts.cpp -I interrupts.hpp -o test1.o 
-./test1.o<<EOF 
-vector_table.txt
-trace1.txt
-execution
-external_files.txt
-EOF
