+++ A2/Part3/test_AjDonald_JayvenLarsen.sh
@@ -1,0 +1,7 @@
+g++ -std=c++11 interrupts_AjDonald_JayvenLarsen.cpp -I interrupts_AjDonald_JayvenLarsen.hpp -o test1.o 
+./test1.o<<EOF 
+vector_table_AjDonald_JayvenLarsen.txt
+trace_AjDonald_JayvenLarsen.txt
+execution_AjDonald_JayvenLarsen
+external_files_AjDonald_JayvenLarsen.txt
+EOF
